+++ UCR.Core/Models/Mapping.cs
@@ -1,8 +1,10 @@
 ﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Xml.Serialization;
 using HidWizards.UCR.Core.Models.Binding;
+using HidWizards.UCR.Core.Models.Subscription;
 
 namespace HidWizards.UCR.Core.Models
 {
@@ -22,6 +24,7 @@ namespace HidWizards.UCR.Core.Models
         internal bool IsShadowMapping { get; set; }
         internal int ShadowDeviceNumber { get; set; }
         internal int PossibleShadowClones => CountPossibleShadowClones();
+        internal FilterState FilterState { get; set; }
 
         private int CountPossibleShadowClones()
         {
@@ -31,7 +34,8 @@ namespace HidWizards.UCR.Core.Models
             {
                 if (!deviceBinding.IsBound) continue;
 
-                usedDeviceConfigurations.Add(Profile.GetDeviceConfiguration(DeviceIoType.Input, deviceBinding.DeviceConfigurationGuid));
+                var deviceConfiguration = Profile.GetDeviceConfiguration(DeviceIoType.Input, deviceBinding.DeviceConfigurationGuid);
+                if (deviceConfiguration != null) usedDeviceConfigurations.Add(deviceConfiguration);
             }
 
             if (usedDeviceConfigurations.Count == 0) return 0;
@@ -83,7 +87,7 @@ namespace HidWizards.UCR.Core.Models
             return result;
         }
 
-        internal void PrepareMapping()
+        internal void PrepareMapping(FilterState filterState)
         {
             InputCache = new List<short>();
             DeviceBindings.ForEach(_ => InputCache.Add(0));
@@ -95,6 +99,9 @@ namespace HidWizards.UCR.Core.Models
                 DeviceBindings[i].Callback = cm.Update;
                 DeviceBindings[i].CurrentValue = 0;
             }
+
+            FilterState = filterState;
+            Plugins.ForEach(p => p.RuntimeMapping = this);
         }
 
         internal Mapping GetOverridenMapping()
@@ -111,6 +118,7 @@ namespace HidWizards.UCR.Core.Models
                 {
                     return mapping;
                 }
+
                 parentProfile = parentProfile?.ParentProfile;
                 if (parentProfile != null) list.AddRange(parentProfile.Mappings);
             }
@@ -122,7 +130,8 @@ namespace HidWizards.UCR.Core.Models
         {
             foreach (var plugin in Plugins)
             {
-                // TODO Surround with Filter check or do pre plugin update
+                if (plugin.IsFiltered()) continue;
+                
                 plugin.Update(InputCache.ToArray());
             }
         }
@@ -182,6 +191,11 @@ namespace HidWizards.UCR.Core.Models
             clonedMapping.Profile = Profile;
             clonedMapping.PostLoad(Profile.Context, Profile);
 
+            foreach (var plugin in clonedMapping.Plugins)
+            {
+                plugin.Filters.ForEach(f => f.Name = Filter.GetShadowName(f.Name, shadowCloneNumber));
+            }
+
             return clonedMapping;
         }
 
@@ -198,15 +212,5 @@ namespace HidWizards.UCR.Core.Models
                 plugin.PostLoad(context, profile);
             }
         }
-
-        internal void InitializeMappings(int amount)
-        {
-            DeviceBindings = new List<DeviceBinding>();
-            for (var i = 0; i < amount; i++)
-            {
-                DeviceBindings.Add(new DeviceBinding(Update, Profile, DeviceIoType.Input));
-            }
-        }
-
     }
 }