+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added Shadow devices
 - Added option to clear bindings
+- Added Filters for plugins
+- `Button to Filter`: Allows toggling of a filter using a button
+- `Axis to Filter`: Allows toggling of a filter using an axis
 
 ### Fixed